refactor array-testing
